+++ examples/t16_global_blackboard.rs
@@ -4,7 +4,7 @@
 //! [tutorial:](https://https://www.behaviortree.dev/docs/tutorial-advanced/tutorial_16_global_blackboard).
 //! [cpp-source:](https://github.com/BehaviorTree/BehaviorTree.CPP/blob/master/examples/t16_global_blackboard.cpp).
 
-use behaviortree::{port::PortRemappings, prelude::*};
+use behaviortree::prelude::*;
 
 const XML: &str = r#"
 <root BTCPP_format="4">
@@ -67,8 +67,7 @@ async fn example() -> BehaviorTreeResult {
 	// create an external blackboard which will survive the tree
 	let mut global_blackboard = SharedBlackboard::default();
 	// BT-Trees blackboard has global blackboard as parent
-	let root_blackboard =
-		SharedBlackboard::with_parent("global", global_blackboard.clone(), PortRemappings::default(), false);
+	let root_blackboard = SharedBlackboard::with_parent("global", global_blackboard.clone(), Remappings::default(), false);
 
 	let mut factory = BehaviorTreeFactory::with_groot2_behaviors()?;