+++ CHANGELOG.md
@@ -26,6 +26,7 @@ especially the [Rust flavour](https://doc.rust-lang.org/cargo/reference/semver.h
 - replaced `expect(SHOULD_NOT_HAPPEN)` with better error handling
 - `SharedQueue` as separate module
 - reorganized factory constructor methods
+- `port::PortRemappings` moved to `blackboard::Remappings`
 
 ### Fixed
 - several `todo!()`'s and `expect(...)`'s